reorganise app
