+++ covey-egui/src/main.rs
@@ -1,7 +1,6 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
 
-use covey_egui::cli;
-use eframe::egui;
+use covey_egui::{App, cli};
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::EnvFilter;
 
@@ -15,12 +14,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .unwrap()
         .add_directive("covey=debug".parse().unwrap());
     tracing_subscriber::fmt().with_env_filter(filter).init();
-
-    let options = eframe::NativeOptions {
-        run_and_return: true,
-        viewport: egui::ViewportBuilder::default().with_inner_size([320.0, 240.0]),
-        ..Default::default()
-    };
 
     // https://github.com/emilk/egui/blob/main/examples/external_eventloop/src/main.rs
     // let mut event_loop = EventLoop::<UserEvent>::with_user_event().build().unwrap();
@@ -36,11 +29,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     tracing::info!("Starting window");
-    eframe::run_native(
-        "covey",
-        options.clone(),
-        Box::new(|_cc| Ok(Box::new(MyApp { rx: rx.clone() }))),
-    )?;
+    App::open(&rx)?;
 
     loop {
         tracing::info!("closed window");
@@ -60,44 +49,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
 
         tracing::info!("Starting window");
-        eframe::run_native(
-            "covey",
-            options.clone(),
-            Box::new(|_cc| Ok(Box::new(MyApp { rx: rx.clone() }))),
-        )?;
+        App::open(&rx)?;
     }
 
     tracing::info!("exiting");
     Ok(())
-}
-
-struct MyApp {
-    rx: cli::Receiver,
-}
-
-impl eframe::App for MyApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
-            match self.rx.try_recv() {
-                Some(cli::Message::Exit) => {
-                    tracing::info!("received exit message");
-                    ui.ctx().send_viewport_cmd(egui::ViewportCommand::Close)
-                }
-                // Trying to open while already open -> do nothing
-                Some(cli::Message::Open) => {}
-                None => {}
-            }
-
-            let label_text =
-                "When this window is closed the next will be opened after a short delay";
-            ui.label(label_text);
-
-            // https://github.com/emilk/egui/issues/3655#issuecomment-3239209608
-            // TODO: try the above
-            if ui.button("Close").clicked() {
-                tracing::info!("Pressed Close button");
-                ui.ctx().send_viewport_cmd(egui::ViewportCommand::Close);
-            }
-        });
-    }
 }
